LPS-55584 Move PortalSessionAuthVerifier into OSGi module
