+++ portal-impl/src/portal.properties
@@ -3654,11 +3654,11 @@
     #
     # PortalSessionAuthVerifier
     #
-    auth.verifier.PortalSessionAuthVerifier.hosts.allowed=
-    auth.verifier.PortalSessionAuthVerifier.urls.includes=\
-        /api/json/*,\
-        /api/jsonws/*,\
-        /c/portal/json_service/*
+    #auth.verifier.PortalSessionAuthVerifier.hosts.allowed=
+    #auth.verifier.PortalSessionAuthVerifier.urls.includes=\
+    #    /api/json/*,\
+    #    /api/jsonws/*,\
+    #    /c/portal/json_service/*
 
     #
     # TunnelingServletAuthVerifier
